LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-inset-animation-expected.txt


PASS view-timeline-inset with one value
PASS view-timeline-inset with two values
PASS view-timeline-inset with em values
PASS view-timeline-inset with percentage values
PASS view-timeline-inset with negative values
FAIL view-timeline-inset with horizontal scroller assert_equals: expected "50" but got "30"
PASS view-timeline-inset with block scroller
FAIL view-timeline-inset with inline scroller assert_equals: expected "50" but got "30"
FAIL view-timeline-inset:auto, block assert_equals: expected "-1" but got "0"
FAIL view-timeline-inset:auto, block, vertical-lr assert_equals: expected "0" but got "15"
FAIL view-timeline-inset:auto, block, vertical-rl assert_equals: expected "0" but got "15"
FAIL view-timeline-inset:auto, inline assert_equals: expected "0" but got "15"
FAIL view-timeline-inset:auto, inline, vertical-rl assert_equals: expected "-1" but got "0"
FAIL view-timeline-inset:auto, inline, vertical-lr assert_equals: expected "-1" but got "0"
FAIL view-timeline-inset:auto, inline, rtl assert_equals: expected "0" but got "15"
FAIL view-timeline-inset:auto, inline, vertical-rl, rtl assert_equals: expected "-1" but got "0"
FAIL view-timeline-inset:auto, inline, vertical-lr, rtl assert_equals: expected "-1" but got "0"
FAIL view-timeline-inset:auto, y assert_equals: expected "-1" but got "0"
FAIL view-timeline-inset:auto, y, vertical-rl assert_equals: expected "-1" but got "0"
FAIL view-timeline-inset:auto, y, vertical-rl, rtl assert_equals: expected "-1" but got "0"
FAIL view-timeline-inset:auto, x assert_equals: expected "0" but got "15"
FAIL view-timeline-inset:auto, x, rtl assert_equals: expected "0" but got "15"
FAIL view-timeline-inset:auto, x, vertical-lr assert_equals: expected "0" but got "15"
FAIL view-timeline-inset:auto, x, vertical-rl assert_equals: expected "0" but got "15"
FAIL view-timeline-inset:auto, mix assert_equals: expected "-1" but got "0"
